+++ test/decorators-duplicate-fixture-name/fixtures.ts
@@ -1,0 +1,25 @@
+import { test as base } from '../../dist';
+import { Fixture, Given } from '../../dist/decorators';
+
+type Fixtures = {
+  todoPage: TodoPage;
+  todoPage2: TodoPage2;
+};
+
+export const test = base.extend<Fixtures>({
+  todoPage: ({}, use) => use(new TodoPage()),
+  todoPage2: ({}, use) => use(new TodoPage2()),
+});
+
+@Fixture<typeof test>('todoPage')
+class TodoPage {
+  @Given('TodoPage: step')
+  async step() {}
+}
+
+// notice the same fixture name
+@Fixture<typeof test>('todoPage')
+class TodoPage2 {
+  @Given('TodoPage2: step')
+  async step() {}
+}