+++ src/utils/index.ts
@@ -1,11 +1,5 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { logger } from './logger';
-
-export function exitWithMessage(...messages: string[]) {
-  logger.error('ERROR:', ...messages.filter(Boolean));
-  process.exit(1);
-}
 
 // See: https://stackoverflow.com/questions/50453640/how-can-i-get-the-value-of-a-symbol-property
 export function getSymbolByName<T extends object>(target: T, name?: string) {