+++ test/error-import-test-from-steps/test.mjs
@@ -1,5 +1,8 @@
 import { test, getTestName, execPlaywrightTestWithError } from '../helpers.mjs';
 
 test(getTestName(import.meta), (t) =>
-  execPlaywrightTestWithError(t.name, `Option "importTestFrom" should point to separate file`),
+  execPlaywrightTestWithError(t.name, [
+    `Option "importTestFrom" should point to a separate file`,
+    /Given, When, Then\)$/, // no other text after error
+  ]),
 );