+++ src/playwright/testTypeImpl.ts
@@ -27,6 +27,25 @@ export function getTestImpl(test: TestTypeCommon) {
 }
 
 /**
+ * Run step with location pointing to Given, When, Then call.
+ */
+// eslint-disable-next-line max-params
+export async function runStepWithCustomLocation(
+  test: TestTypeCommon,
+  stepText: string,
+  location: Location,
+  body: () => unknown,
+) {
+  const testInfo = test.info();
+
+  // See: https://github.com/microsoft/playwright/blob/main/packages/playwright/src/common/testType.ts#L221
+  // @ts-expect-error _runAsStep is hidden from public
+  return testInfo._runAsStep({ category: 'test.step', title: stepText, location }, async () => {
+    return await body();
+  });
+}
+
+/**
  * Returns true if all fixtures of parent test found in child test.
  */
 export function isParentChildTest(parent: TestTypeCommon, child: TestTypeCommon) {