+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## dev
-* i18n: Generate scenario outlines correctly [#60](https://github.com/vitalets/playwright-bdd/issues/60)).
+* i18n: Generate scenario outlines correctly [#60](https://github.com/vitalets/playwright-bdd/issues/60).
+* Check for duplicate fixture names [#52](https://github.com/vitalets/playwright-bdd/issues/52)
 
 ## 5.3.0
 * Add support for Playwright `1.38`.