+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## dev
 * i18n: Generate scenario outlines correctly [#60](https://github.com/vitalets/playwright-bdd/issues/60).
 * Check for duplicate fixture names [#52](https://github.com/vitalets/playwright-bdd/issues/52)
+* Fix flushing logs for several projects [#59](https://github.com/vitalets/playwright-bdd/issues/59)
 
 ## 5.3.0
 * Add support for Playwright `1.38`.