+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Check for duplicate fixture names [#52](https://github.com/vitalets/playwright-bdd/issues/52)
 * Fix flushing logs for several projects [#59](https://github.com/vitalets/playwright-bdd/issues/59)
 * Support Playwright `1.39`.
+* Docs: add search.
 
 ## 5.3.0
 * Add support for Playwright `1.38`.