+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## dev
+## 5.4.0
 * i18n: Generate scenario outlines correctly [#60](https://github.com/vitalets/playwright-bdd/issues/60).
 * Check for duplicate fixture names [#52](https://github.com/vitalets/playwright-bdd/issues/52)
 * Fix flushing logs for several projects [#59](https://github.com/vitalets/playwright-bdd/issues/59)